+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - Fix bug with allUSDT swaps resulting in extreme price impact due to config and filtering issues.
 - Fix bug of overapplying quote denom scaling factor in pool liquidity pricer.
 - Pool liquidity capitalization
+- Make alloyed transmuter pools receive more rank in routing
 
 ## v25.1.1